requirements.txt
suds-jurko>=0.6
python-social-auth>=0.2.1
six>=1.7.2